requirements.txt
fastapi
uvicorn
python-dotenv
requests
mysql-connector-python